+++ lib/app/widgets/home_search_dialog.dart
@@ -142,6 +142,7 @@ class HomeSearchDialogState extends State<HomeSearchDialog> {
                   ),
                 ),
                 IconButton(
+                  tooltip: AppLocalizations.of(context)!.close,
                   icon: Icon(Icons.close, color: Colors.white),
                   onPressed: () => Navigator.pop(context),
                 ),